backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
ollama==0.4.4
tavily-python==0.5.0
openai==1.57.4
httpx>=0.27.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
phonenumbers>=8.13.0
dnspython>=2.6.0
